Split tile_based_layout into a hierarchy of templatized layouts: tile_based_layout, clocked_layout, gate_level_layout
